refactor: 도메인 모듈 생성
