+++ domain/src/main/java/com/ticketlounge/domain/member/Member.java
@@ -1,4 +1,4 @@
-package com.ticketexchange.member.domain;
+package com.ticketlounge.domain.member;
 
 public class Member {
