+++ domain/src/main/java/com/ticketlounge/domain/member/MemberCreateEvent.java
@@ -1,4 +1,4 @@
-package com.ticketexchange.member.domain;
+package com.ticketlounge.domain.member;
 
 public record MemberCreateEvent(Long memberId) {
 }
