+++ domain/src/main/java/com/ticketlounge/domain/product/EarnedProduct.java
@@ -1,4 +1,4 @@
-package com.ticketexchange.product.domain;
+package com.ticketlounge.domain.product;
 
 import java.time.LocalDateTime;
