+++ domain/src/test/java/com/ticketexchange/product/ProbabilityTest.java
@@ -1,4 +1,4 @@
-package com.ticketexchange.product.domain.vo;
+package com.ticketexchange.product;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -7,7 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import org.junit.jupiter.api.Test;
 
-import com.ticketexchange.product.domain.Probability;
+import com.ticketlounge.domain.product.Probability;
 
 class ProbabilityTest {
